Backend-Part/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
assemblyai==0.46.0
attrs==25.4.0
av==15.1.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.3.0
fastapi==0.123.5
flatbuffers==25.9.23
frozenlist==1.8.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-speech==2.34.0
google-cloud-texttospeech==2.33.0
google-genai==1.52.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
jiter==0.12.0
langdetect==1.0.9
livekit==1.0.19
livekit-agents==1.2.18
livekit-api==1.0.7
livekit-plugins-google==1.2.18
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-silero==1.2.18
livekit-protocol==1.1.0
markdown-it-py==4.0.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
nest-asyncio==1.6.0
numpy==2.3.5
openai==2.8.1
packaging==25.0
pillow==12.0.0
prometheus_client==0.23.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyreadline3==3.5.4
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tqdm==4.67.1
typer==0.20.0
types-protobuf==6.32.1.20251105
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
